refactor: UI - Extract Header
